+++ app/Livewire/Profile/Settings.php
@@ -3,6 +3,7 @@
 namespace App\Livewire\Profile;
 
 use App\Models\Configuration;
+use App\Models\ConfigurationKey;
 use App\Models\User;
 use Livewire\Attributes\Renderless;
 use Livewire\Component;
@@ -12,13 +13,13 @@ class Settings extends Component
     #[Renderless]
     public function dashboardButtonChangedBirthdayList(bool $enabled): void
     {
-        Configuration::storeBool('dashboardButtons_birthdayList', $enabled, auth()->user());
+        Configuration::storeBool(ConfigurationKey::DASHBOARD_BTN_BIRTHDAY_LIST, $enabled, auth()->user());
     }
 
     #[Renderless]
     public function dashboardButtonChangedImportMembers(bool $enabled): void
     {
-        Configuration::storeBool('dashboardButtons_importMembers', $enabled, auth()->user());
+        Configuration::storeBool(ConfigurationKey::DASHBOARD_BTN_IMPORT_MEMBERS, $enabled, auth()->user());
     }
 
     public function render(): \Illuminate\Contracts\View\View|\Illuminate\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\Foundation\Application