+++ resources/views/livewire/members/member-overview.blade.php
@@ -52,12 +52,12 @@
                     @php
                         $rowBg = "bg-lime-200";
                         if($member->entrance_date === null || $member->birthday === null) {
-                            $rowBg = "bg-red-200";
+                            $rowBg = "bg-lime-200 text-red-700";
                         } elseif($member->entrance_date > now() ||
                          ($member->leaving_date && $member->leaving_date < now())) {
-                            $rowBg = "bg-gray-300";
+                            $rowBg = "bg-gray-300 text-gray-600";
                         } elseif($member->paused) {
-                            $rowBg = "bg-sky-200";
+                            $rowBg = "bg-gray-300";
                         }
                     @endphp
                     <tr class="[&:nth-child(2n)]:bg-opacity-50 {{$rowBg}}">
@@ -119,7 +119,7 @@
                                 <x-button-link class="w-full"
                                                href="{{route('member.list.excel', $memberFilter->toParameterArray())}}"
                                                @click="open=false"
-                                               title="Download birthday list as CSV file">{{ __('Excel File') }}</x-button-link>
+                                               title="Download birthday list as Excel file">{{ __('Excel File') }}</x-button-link>
                             </div>
                         </div>
                     </div>